LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/stale-while-revalidate/stale-image-expected.txt



FAIL Cache returns stale resource assert_equals: (revalidated version loaded) expected 256 but got 16
